+++ ares/__init__.py
@@ -1,18 +1,22 @@
 import logging
 
-import gym
 from gym.envs.registration import register
 
-from ares import environment, utils
+from ares import attacker, defender, scenario, utils
+from ares.environment import AresEnvironment
+from ares.utils import load_config
 
 logger = logging.getLogger(__name__)
 
-register(id="AresEnv-v0", entry_point="ares.environment:AresEnv")
+register(id="AresEnvironment-v0", entry_point="ares.environment:AresEnvironment")
 
 
-def create_env(config: dict, device) -> environment.AresEnv:
-    defender_agent = utils.get_defender_agent(config, device)
-    attacker_agent = utils.get_attacker_agent(config)
-    execution_scenario = utils.get_execution_scenario(config)
-    env = gym.make("AresEnv-v0", attacker=attacker_agent, defender=defender_agent, scenario=execution_scenario)
+def construct(config: dict, device) -> AresEnvironment:
+    attacker_agent = attacker.get_attacker_agent(config)
+    defender_agent = defender.get_defender_agent(config, device)
+    execution_scenario = scenario.get_evaluation_scenario(config)
+    env = AresEnvironment(attacker_agent, defender_agent, execution_scenario)
     return env
+
+
+__all__ = ["utils", "attacker", "defender", "scenario", "load_config", "construct"]
