+++ ares/attacker/__init__.py
@@ -1,3 +1,3 @@
-from .attacker_agent import AttackConfig, AttackerAgent
+from .attacker_agent import AttackConfig, AttackerAgent, get_attacker_agent
 
-__all__ = ["AttackConfig", "AttackerAgent"]
+__all__ = ["AttackConfig", "AttackerAgent", "get_attacker_agent"]
