+++ ares/defender/__init__.py
@@ -1,4 +1,5 @@
-from .defender_agent import DefenderAgent
+from .classifiers import get_classifier
+from .defender_agent import DefenderAgent, get_defender_agent
 from .detector import Detector, DetectorConfig, get_detector
 
-__all__ = ["DefenderAgent", "Detector", "DetectorConfig", "get_detector"]
+__all__ = ["DefenderAgent", "Detector", "DetectorConfig", "get_detector", "get_classifier", "get_defender_agent"]
