+++ ares/defender/defender_agent.py
@@ -3,8 +3,10 @@ from typing import List, Optional
 from art.estimators.classification import PyTorchClassifier
 from gym import spaces
 import numpy as np
+import torch
 
-from ares.defender.detector import Detector
+from ares.defender.classifiers import get_classifier, get_torch_model
+from ares.defender.detector import Detector, DetectorConfig, get_detector
 
 
 class DefenderAgent:
@@ -30,3 +32,35 @@ class DefenderAgent:
         if self.detector:
             return self.detector.detect(x)
         return False
+
+
+def get_defender_agent(config: dict, device: torch.device) -> DefenderAgent:
+    defender_models = config["defender"]["models"]
+    classifiers = []
+
+    for defender_model in defender_models:
+        model_file = defender_model["model_file"]
+        model_name = defender_model["model_name"]
+        model_params = defender_model["model_params"]
+        model_state = defender_model["model_state"]
+        classifier_type = defender_model["classifier_type"]
+        classifier_params = defender_model["classifier_params"]
+        model = get_torch_model(model_file, model_name, model_params, model_state, device)
+        classifier = get_classifier(model, classifier_type, classifier_params, device)
+        classifiers.append(classifier)
+
+    probs = config["defender"].get("probabilities", None)
+    if probs:
+        probs = np.array(probs) / np.sum(probs)
+    else:
+        probs = np.ones(len(classifiers)) / len(classifiers)
+
+    detector_args = config["defender"].get("detector", None)
+    if detector_args:
+        detector_config = DetectorConfig(**detector_args)
+        detector = get_detector(detector_config)
+    else:
+        detector = None
+
+    defender_agent = DefenderAgent(classifiers, probs, detector)
+    return defender_agent
