+++ ares/scenario/__init__.py
@@ -1,0 +1,3 @@
+from .evaluation_scenario import EvaluationScenario, get_evaluation_scenario
+
+__all__ = ["EvaluationScenario", "get_evaluation_scenario"]
