+++ setup.py
@@ -6,22 +6,34 @@ with open("README.md", "r") as f:
 install_requires = [
     "numpy",
     "scipy",
-    "Pillow",
-    "tqdm",
-    "torch",
-    "torchvision",
     "gym",
-    "adversarial-robustness-toolbox",
 ]
+dev_requires = ["pytest", "tox"]
+sklearn_requires = ["scikit-learn"]
+torch_requires = ["torch", "torchvision"]
+tf_requires = ["tensorflow"]
+art_requires = ["adversarial-robustness-toolbox"]
+all_requires = dev_requires + sklearn_requires + torch_requires + tf_requires + art_requires
 
 setup(
     name="ares",
-    version="0.0.1",
+    version="1.0.0",
     description="A System-Oriented Wargame Framework for Adversarial ML",
     long_description=long_description,
     long_description_content_type="text/markdown",
     license="MIT",
     install_requires=install_requires,
-    extras_require={"dev": ["pytest", "tox"]},
+    extras_require={
+        "dev": dev_requires,
+        "scikit-learn": sklearn_requires,
+        "sklearn": sklearn_requires,
+        "pytorch": torch_requires,
+        "torch": torch_requires,
+        "tensorflow": tf_requires,
+        "tf": tf_requires,
+        "adversarial-robustness-toolbox": art_requires,
+        "art": art_requires,
+        "all": all_requires,
+    },
     packages=find_packages(),
 )
